chore: Removed not needed files
