+++ src/backend/settings/analze_settings.hpp
@@ -22,7 +22,6 @@
 #include <string>
 #include "backend/enums/enums_file_endians.hpp"
 #include "backend/settings/program/program_meta.hpp"
-#include "backend/settings/results_settings/results_template.hpp"
 #include "pipeline/pipeline.hpp"
 #include "project_settings/project_settings.hpp"
 #include "results_settings/results_settings.hpp"
@@ -51,6 +50,7 @@ public:
   auto getImageChannelsUsedForMeasurement() const -> std::map<enums::ClassId, std::set<int32_t>>;
 
   auto checkForErrors() const -> std::vector<std::pair<std::string, SettingParserLog_t>>;
+  auto toResultsSettings() const -> ResultsSettings;
 
 private:
   std::string configSchema = "https://imagec.org/schemas/v1/analyze-settings.json";
