+++ src/backend/settings/project_settings/project_class.hpp
@@ -14,8 +14,8 @@
 
 #include <set>
 #include <vector>
+#include "backend/enums/enum_measurements.hpp"
 #include "backend/enums/enums_classes.hpp"
-#include "backend/settings/results_settings/results_template.hpp"
 #include "backend/settings/setting.hpp"
 #include <nlohmann/json.hpp>
 
@@ -55,6 +55,31 @@ static inline const std::vector<std::string> COLORS = {
 
     // Gray
     "#BFBFBF"    // -Light gray
+};
+
+///
+/// \class      ResultsTemplate
+/// \author     Joachim Danmayr
+/// \brief      Template for results
+///
+struct ResultsTemplate
+{
+  enums::Measurement measureChannel = enums::Measurement::NONE;
+  enums::Stats stats                = enums::Stats::AVG;
+  // int32_t crossChannelStacksC              = -1;
+  // joda::enums::ClassId intersectingChannel = joda::enums::ClassId::NONE;
+  // int32_t zStack                           = 0;
+  // int32_t tStack                           = 0;
+
+  NLOHMANN_DEFINE_TYPE_INTRUSIVE_WITH_DEFAULT(ResultsTemplate, measureChannel, stats);
+
+  void check() const
+  {
+  }
+  // We don't want to do a error check for the history
+  void getErrorLogRecursive(SettingParserLog_t &settingsParserLog) const
+  {
+  }
 };
 
 struct Class
